fixed all code in home
